Moved random generator to main scope.
